"fix : MainActivity 클린하게하기"
